custom and colors update 1
