--- lib/core/widgets/undefined_route_page.dart
@@ -1,19 +1,0 @@
-import 'package:flutter/material.dart';
-
-class UndefinedRoutePage extends StatelessWidget {
-  const UndefinedRoutePage({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text("No Route Found"),
-      ),
-      body: Center(
-        child: Text("wrong route page"),
-      ),
-    );
-  }
-}
